Remove Backdoor: Authentication bypass via extract($_REQUEST["fx29shcook"]) & $auth['login']
Passing user input ($_REQUEST["fx29shcook"]) to the PHP extract function allows a user to overwrite any variable using an array.

Via normal usage of this shell, a user setting the $auth['login'] and $auth['pass'] variables would expect the shell to be protected via that authentication.

However due to the code extract($_REQUEST["fx29shcook"]), if a user appends ?fx29shcook[auth]=0 to the url the variable $auth['login'] will be overwritten, with the authentication credentials replaced by a 0 (false) value.

Later in the code when an authentication check is performed - if (!empty($auth["login"])) - the authentication is bypassed exactly the same as if the $auth['login'] variable was not set.