Remove Backdoor: Authentication bypass via extract($_REQUEST["N3tshcook"]) & $login
Passing user input ($_REQUEST["N3tshcook"]) to the PHP extract function allows a user to overwrite any variable using an array.

Via normal usage of this shell, a user setting the $login and $pass variables would expect the shell to be protected via that authentication.

However due to the code extract($_REQUEST["N3tshcook"]), if a user appends ?N3tshcook[login]=0 to the url the variable $login will be overwritten, with the authentication credentials replaced by a 0 (false) value.

Later in the code when an authentication check is performed - if (!empty($login)) - the authentication is bypassed exactly the same as if the $login variable was not set.